requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
httpx==0.27.2
python-multipart==0.0.9
pydantic==2.9.2
jinja2==3.1.4
python-dotenv==1.0.1
dnspython==2.7.0